Rename hbaseUtils into hbase_utils
