+++ bin/daily_stats.py
@@ -23,7 +23,7 @@ import pandas as pd
 import pyspark.sql.functions as F
 
 from fink_broker.sparkUtils import init_sparksession
-from fink_broker.hbaseUtils import push_to_hbase
+from fink_broker.hbase_utils import push_to_hbase
 from fink_broker.parser import getargs
 from fink_broker.loggingUtils import get_fink_logger, inspect_application
