+++ bin/index_archival.py
@@ -33,12 +33,12 @@ from pyspark.sql.functions import pandas_udf, PandasUDFType
 
 from fink_broker.parser import getargs
 from fink_broker.science import ang2pix
-from fink_broker.hbaseUtils import push_to_hbase, add_row_key
-from fink_broker.hbaseUtils import assign_column_family_names
-from fink_broker.hbaseUtils import load_ztf_index_cols
-from fink_broker.hbaseUtils import load_ztf_crossmatch_cols
-from fink_broker.hbaseUtils import select_relevant_columns
-from fink_broker.hbaseUtils import bring_to_current_schema
+from fink_broker.hbase_utils import push_to_hbase, add_row_key
+from fink_broker.hbase_utils import assign_column_family_names
+from fink_broker.hbase_utils import load_ztf_index_cols
+from fink_broker.hbase_utils import load_ztf_crossmatch_cols
+from fink_broker.hbase_utils import select_relevant_columns
+from fink_broker.hbase_utils import bring_to_current_schema
 from fink_broker.sparkUtils import init_sparksession, load_parquet_files
 from fink_broker.loggingUtils import get_fink_logger, inspect_application
